medicaltutordjproject/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
distro==1.9.0
Django==5.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
Markdown==3.0.1
MarkdownSuperscript==2.1.1
openai==1.70.0
pillow==11.1.0
psycopg2-binary==2.9.10
pydantic==2.11.2
pydantic_core==2.33.1
pygiftparser==1.1
python-dotenv==1.0.0
setuptools==78.1.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
whitenoise==6.9.0
yattag==1.16.1
